requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
executing==1.2.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-language==2.9.0
googleapis-common-protos==1.59.0
googletrans==3.0.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
openai==0.27.4
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
whitenoise==6.3.0
yarl==1.8.2
